pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=23.1.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.7",
    "pyjwt>=2.10.1",
    "redis>=5.2.1",
    "sqlalchemy[asyncio]>=2.0.37",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.14.1",
    "pytest-asyncio>=0.25.3",
]
